fixed the article toggle
The issue might be due to the fact that the openArticles array is not reactive. When you call appStore.updateOpenArticles, it updates the openArticles array in the store, but the openArticles constant in your component still refers to the old array.

To fix this, you can make openArticles a computed property that always returns the current value of appStore.openArticles